requirements.txt
httpx==0.24.1
pydantic-settings==2.0b1
python-telegram-bot==20.3
